// src/app/models/ContactUs.ts
export class ContactUs {
  id: number | undefined;
  name: string | undefined;
  email: string | undefined;
  howCanWeHelpYou: string | undefined;
  attachment_1: string | undefined;
  attachment_2: string | undefined;
  attachment_3: string | undefined;
  attachment_4: string | undefined;
  attachment_5: string | undefined;
}